requirements.txt
pyyaml
numpy
torch
torchvision
torchaudio
ultralytics
pydantic
opencv-python
easyocr
picamera2
